cleaned up code and requested changes
